@@ -66,10 +66,10 @@
   ],
   "devDependencies": {
     "@changesets/changelog-github": "0.2.7",
-    "@changesets/cli": "2.11.0",
+    "@changesets/cli": "2.11.2",
     "api-documenter-me": "0.1.0",
     "api-extractor-me": "0.1.0",
-    "@types/chai": "4.2.13",
+    "@types/chai": "4.2.14",
     "@types/chai-as-promised": "7.1.3",
     "@types/child-process-promise": "2.2.1",
     "@types/clone": "2.1.0",
@@ -77,18 +77,18 @@
     "@types/listr": "0.14.2",
     "@types/long": "4.0.1",
     "@types/mocha": "7.0.2",
-    "@types/mz": "2.7.1",
-    "@types/node": "12.12.67",
+    "@types/mz": "2.7.2",
+    "@types/node": "12.19.3",
     "@types/sinon": "9.0.8",
     "@types/sinon-chai": "3.2.5",
     "@types/tmp": "0.2.0",
-    "@types/yargs": "15.0.8",
-    "@typescript-eslint/eslint-plugin": "4.4.1",
-    "@typescript-eslint/eslint-plugin-tslint": "4.4.1",
-    "@typescript-eslint/parser": "4.4.1",
+    "@types/yargs": "15.0.9",
+    "@typescript-eslint/eslint-plugin": "4.6.1",
+    "@typescript-eslint/eslint-plugin-tslint": "4.6.1",
+    "@typescript-eslint/parser": "4.6.1",
     "babel-loader": "8.1.0",
-    "@babel/core": "7.11.6",
-    "@babel/preset-env": "7.11.5",
+    "@babel/core": "7.12.3",
+    "@babel/preset-env": "7.12.1",
     "@babel/plugin-transform-modules-commonjs": "7.12.1",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
@@ -98,12 +98,12 @@
     "coveralls": "3.1.0",
     "del": "6.0.0",
     "dependency-graph": "0.9.0",
-    "eslint": "7.11.0",
+    "eslint": "7.12.1",
     "eslint-plugin-import": "2.22.1",
     "express": "4.17.1",
     "find-free-port": "2.0.0",
     "firebase-functions": "3.11.0",
-    "firebase-tools": "8.12.1",
+    "firebase-tools": "8.15.0",
     "git-rev-sync": "3.0.1",
     "glob": "7.1.6",
     "http-server": "0.12.3",
@@ -116,7 +116,7 @@
     "karma-chrome-launcher": "3.1.0",
     "karma-cli": "2.0.0",
     "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-firefox-launcher": "2.0.0",
+    "karma-firefox-launcher": "2.1.0",
     "karma-mocha": "2.0.1",
     "karma-mocha-reporter": "2.2.5",
     "karma-safari-launcher": "1.0.0",
@@ -143,18 +143,18 @@
     "rxjs": "6.6.3",
     "semver": "7.3.2",
     "simple-git": "2.21.0",
-    "sinon": "9.2.0",
+    "sinon": "9.2.1",
     "sinon-chai": "3.5.0",
-    "source-map-loader": "1.1.1",
-    "terser": "5.3.5",
-    "ts-loader": "8.0.5",
+    "source-map-loader": "1.1.2",
+    "terser": "5.3.8",
+    "ts-loader": "8.0.9",
     "ts-node": "9.0.0",
     "tslint": "6.1.3",
     "typedoc": "0.16.11",
     "typescript": "4.0.5",
     "watch": "1.0.2",
     "webpack": "4.44.2",
-    "yargs": "16.0.3"
+    "yargs": "16.1.0"
   },
   "husky": {
     "hooks": {
