@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2017 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { assert } from 'chai';
+import { expect } from 'chai';
 import { FirebaseApp } from '@firebase/app-types';
 import { StringFormat } from '../../src/implementation/string';
 import { Headers } from '../../src/implementation/xhrio';
 import { Metadata } from '../../src/metadata';
-import { Reference } from '../../src/reference';
-import { StorageService } from '../../src/service';
+import {
+  Reference,
+  uploadString,
+  uploadBytesResumable,
+  deleteObject,
+  list,
+  getMetadata,
+  updateMetadata,
+  getDownloadURL
+} from '../../src/reference';
+import { StorageService, ref } from '../../src/service';
 import * as testShared from './testshared';
 import { SendHook, TestingXhrIo } from './xhrio';
 import { DEFAULT_HOST } from '../../src/implementation/constants';
@@ -38,7 +47,7 @@ function makeFakeService(
 
 function makeStorage(url: string): Reference {
   const service = new StorageService(
-    null,
+    {} as FirebaseApp,
     testShared.emptyAuthProvider,
     testShared.makePool(null)
   );
@@ -50,28 +59,28 @@ describe('Firebase Storage > Reference', () => {
   const child = makeStorage('gs://test-bucket/hello');
   describe('Path constructor', () => {
     it('root', () => {
-      assert.equal(root.toString(), 'gs://test-bucket/');
+      expect(root.toString()).to.equal('gs://test-bucket/');
     });
     it('keeps characters after ? on a gs:// string', () => {
       const s = makeStorage('gs://test-bucket/this/ismyobject?hello');
-      assert.equal(s.toString(), 'gs://test-bucket/this/ismyobject?hello');
+      expect(s.toString()).to.equal('gs://test-bucket/this/ismyobject?hello');
     });
     it("doesn't URL-decode on a gs:// string", () => {
       const s = makeStorage('gs://test-bucket/%3F');
-      assert.equal(s.toString(), 'gs://test-bucket/%3F');
+      expect(s.toString()).to.equal('gs://test-bucket/%3F');
     });
     it('ignores URL params and fragments on an http URL', () => {
       const s = makeStorage(
         `http://${DEFAULT_HOST}/v0/b/test-bucket/o/my/object.txt` +
           '?ignoreme#please'
       );
-      assert.equal(s.toString(), 'gs://test-bucket/my/object.txt');
+      expect(s.toString()).to.equal('gs://test-bucket/my/object.txt');
     });
     it('URL-decodes and ignores fragment on an http URL', () => {
       const s = makeStorage(
         `http://${DEFAULT_HOST}/v0/b/test-bucket/o/%3F?ignore`
       );
-      assert.equal(s.toString(), 'gs://test-bucket/?');
+      expect(s.toString()).to.equal('gs://test-bucket/?');
     });
 
     it('ignores URL params and fragments on an https URL', () => {
@@ -79,95 +88,93 @@ describe('Firebase Storage > Reference', () => {
         `https://${DEFAULT_HOST}/v0/b/test-bucket/o/my/object.txt` +
           '?ignoreme#please'
       );
-      assert.equal(s.toString(), 'gs://test-bucket/my/object.txt');
+      expect(s.toString()).to.equal('gs://test-bucket/my/object.txt');
     });
 
     it('URL-decodes and ignores fragment on an https URL', () => {
       const s = makeStorage(
         `https://${DEFAULT_HOST}/v0/b/test-bucket/o/%3F?ignore`
       );
-      assert.equal(s.toString(), 'gs://test-bucket/?');
+      expect(s.toString()).to.equal('gs://test-bucket/?');
     });
   });
 
   describe('toString', () => {
     it("Doesn't add trailing slash", () => {
       const s = makeStorage('gs://test-bucket/foo');
-      assert.equal(s.toString(), 'gs://test-bucket/foo');
+      expect(s.toString()).to.equal('gs://test-bucket/foo');
     });
     it('Strips trailing slash', () => {
       const s = makeStorage('gs://test-bucket/foo/');
-      assert.equal(s.toString(), 'gs://test-bucket/foo');
+      expect(s.toString()).to.equal('gs://test-bucket/foo');
     });
   });
 
-  describe('parent', () => {
+  describe('parentReference', () => {
     it('Returns null at root', () => {
-      assert.isNull(root.parent);
+      expect(root.parent).to.be.null;
     });
     it('Returns root one level down', () => {
-      assert.equal(child.parent!.toString(), 'gs://test-bucket/');
+      expect(child.parent!.toString()).to.equal('gs://test-bucket/');
     });
     it('Works correctly with empty levels', () => {
       const s = makeStorage('gs://test-bucket/a///');
-      assert.equal(s.parent!.toString(), 'gs://test-bucket/a/');
+      expect(s.parent!.toString()).to.equal('gs://test-bucket/a/');
     });
   });
 
   describe('root', () => {
     it('Returns self at root', () => {
-      assert.equal(root.root.toString(), 'gs://test-bucket/');
+      expect(root.root.toString()).to.equal('gs://test-bucket/');
     });
 
     it('Returns root multiple levels down', () => {
       const s = makeStorage('gs://test-bucket/a/b/c/d');
-      assert.equal(s.root.toString(), 'gs://test-bucket/');
+      expect(s.root.toString()).to.equal('gs://test-bucket/');
     });
   });
 
   describe('bucket', () => {
     it('Returns bucket name', () => {
-      assert.equal(root.bucket, 'test-bucket');
+      expect(root.bucket).to.equal('test-bucket');
     });
   });
 
   describe('fullPath', () => {
     it('Returns full path without leading slash', () => {
       const s = makeStorage('gs://test-bucket/full/path');
-      assert.equal(s.fullPath, 'full/path');
+      expect(s.fullPath).to.equal('full/path');
     });
   });
 
   describe('name', () => {
     it('Works at top level', () => {
       const s = makeStorage('gs://test-bucket/toplevel.txt');
-      assert.equal(s.name, 'toplevel.txt');
+      expect(s.name).to.equal('toplevel.txt');
     });
 
     it('Works at not the top level', () => {
       const s = makeStorage('gs://test-bucket/not/toplevel.txt');
-      assert.equal('toplevel.txt', s.name);
+      expect(s.name).to.equal('toplevel.txt');
     });
   });
 
-  describe('child', () => {
+  describe('get child with ref()', () => {
     it('works with a simple string', () => {
-      assert.equal(root.child('a').toString(), 'gs://test-bucket/a');
+      expect(ref(root, 'a').toString()).to.equal('gs://test-bucket/a');
     });
     it('drops a trailing slash', () => {
-      assert.equal(root.child('ab/').toString(), 'gs://test-bucket/ab');
+      expect(ref(root, 'ab/').toString()).to.equal('gs://test-bucket/ab');
     });
     it('compresses repeated slashes', () => {
-      assert.equal(
-        root.child('//a///b/////').toString(),
+      expect(ref(root, '//a///b/////').toString()).to.equal(
         'gs://test-bucket/a/b'
       );
     });
     it('works chained multiple times with leading slashes', () => {
-      assert.equal(
-        root.child('a').child('/b').child('c').child('d/e').toString(),
-        'gs://test-bucket/a/b/c/d/e'
-      );
+      expect(
+        ref(ref(ref(ref(root, 'a'), '/b'), 'c'), 'd/e').toString()
+      ).to.equal('gs://test-bucket/a/b/c/d/e');
     });
   });
 
@@ -179,8 +186,8 @@ describe('Firebase Storage > Reference', () => {
       body?: ArrayBufferView | Blob | string | null,
       headers?: Headers
     ): void {
-      assert.isDefined(headers);
-      assert.isUndefined(headers!['Authorization']);
+      expect(headers).to.not.be.undefined;
+      expect(headers!['Authorization']).to.be.undefined;
       done();
     }
 
@@ -189,8 +196,8 @@ describe('Firebase Storage > Reference', () => {
       testShared.emptyAuthProvider,
       newSend
     );
-    const ref = service.refFromURL('gs://test-bucket');
-    ref.child('foo').getMetadata();
+    const reference = ref(service, 'gs://test-bucket');
+    getMetadata(ref(reference, 'foo'));
   });
 
   it('Works if the user logs in before creating the storage reference', done => {
@@ -202,9 +209,8 @@ describe('Firebase Storage > Reference', () => {
       body?: ArrayBufferView | Blob | string | null,
       headers?: Headers
     ): void {
-      assert.isDefined(headers);
-      assert.equal(
-        headers!['Authorization'],
+      expect(headers).to.not.be.undefined;
+      expect(headers!['Authorization']).to.equal(
         'Firebase ' + testShared.authToken
       );
       done();
@@ -215,168 +221,84 @@ describe('Firebase Storage > Reference', () => {
       testShared.fakeAuthProvider,
       newSend
     );
-    const ref = service.refFromURL('gs://test-bucket');
-    ref.child('foo').getMetadata();
+    const reference = ref(service, 'gs://test-bucket');
+    getMetadata(ref(reference, 'foo'));
   });
 
-  describe('putString', () => {
+  describe('uploadString', () => {
     it('Uses metadata.contentType for RAW format', () => {
       // Regression test for b/30989476
-      const task = child.putString('hello', StringFormat.RAW, {
+      const task = uploadString(child, 'hello', StringFormat.RAW, {
         contentType: 'lol/wut'
       } as Metadata);
-      assert.equal(task.snapshot.metadata!.contentType, 'lol/wut');
+      expect(task.snapshot.metadata!.contentType).to.equal('lol/wut');
       task.cancel();
     });
     it('Uses embedded content type in DATA_URL format', () => {
-      const task = child.putString(
+      const task = uploadString(
+        child,
         'data:lol/wat;base64,aaaa',
         StringFormat.DATA_URL
       );
-      assert.equal(task.snapshot.metadata!.contentType, 'lol/wat');
+      expect(task.snapshot.metadata!.contentType).to.equal('lol/wat');
       task.cancel();
     });
     it('Lets metadata.contentType override embedded content type in DATA_URL format', () => {
-      const task = child.putString(
+      const task = uploadString(
+        child,
         'data:ignore/me;base64,aaaa',
         StringFormat.DATA_URL,
         { contentType: 'tomato/soup' } as Metadata
       );
-      assert.equal(task.snapshot.metadata!.contentType, 'tomato/soup');
+      expect(task.snapshot.metadata!.contentType).to.equal('tomato/soup');
       task.cancel();
     });
   });
 
   describe('Argument verification', () => {
     describe('list', () => {
-      it('throws on invalid maxResults', () => {
-        testShared.assertThrows(
-          testShared.bind(child.list, child, { maxResults: 0 }),
+      it('throws on invalid maxResults', async () => {
+        await expect(list(child, { maxResults: 0 })).to.be.rejectedWith(
           'storage/invalid-argument'
         );
-        testShared.assertThrows(
-          testShared.bind(child.list, child, { maxResults: -4 }),
+        await expect(list(child, { maxResults: -4 })).to.be.rejectedWith(
           'storage/invalid-argument'
         );
-        testShared.assertThrows(
-          testShared.bind(child.list, child, { maxResults: 1001 }),
+        await expect(list(child, { maxResults: 1001 })).to.be.rejectedWith(
           'storage/invalid-argument'
         );
-      });
-    });
-  });
-
-  describe('non-root operations', () => {
-    it("put doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.put(new Blob(['a']));
-        child.put(new Uint8Array(10));
-        child.put(new ArrayBuffer(10));
-      });
-    });
-    it("putString doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.putString('raw', StringFormat.RAW);
-        child.putString('aaaa', StringFormat.BASE64);
-        child.putString('aaaa', StringFormat.BASE64URL);
-        child.putString(
-          'data:application/octet-stream;base64,aaaa',
-          StringFormat.DATA_URL
-        );
-      });
-    });
-    it("delete doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.delete();
-      });
-    });
-    it("getMetadata doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.getMetadata();
-      });
-    });
-    it("listAll doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.listAll();
-      });
-    });
-    it("list doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.list();
-      });
-      assert.doesNotThrow(() => {
-        child.list({ pageToken: 'xxx', maxResults: 4 });
-      });
-      assert.doesNotThrow(() => {
-        child.list({ pageToken: 'xxx' });
-      });
-      assert.doesNotThrow(() => {
-        child.list({ maxResults: 4 });
-      });
-      assert.doesNotThrow(() => {
-        child.list({ maxResults: 4, pageToken: null });
-      });
-    });
-    it("updateMetadata doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.updateMetadata({} as Metadata);
-      });
-    });
-    it("getDownloadURL doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        child.getDownloadURL();
       });
     });
   });
 
   describe('root operations', () => {
-    it('put throws', () => {
-      testShared.assertThrows(
-        root.put.bind(root, new Blob(['a'])),
+    it('uploadBytesResumable throws', () => {
+      expect(() => uploadBytesResumable(root, new Blob(['a']))).to.throw(
         'storage/invalid-root-operation'
       );
     });
-    it('putString throws', () => {
-      testShared.assertThrows(
-        root.putString.bind(root, 'raw', StringFormat.RAW),
+    it('uploadString throws', () => {
+      expect(() => uploadString(root, 'raw', StringFormat.RAW)).to.throw(
         'storage/invalid-root-operation'
       );
     });
-    it('delete throws', () => {
-      testShared.assertThrows(
-        root.delete.bind(root),
+    it('deleteObject throws', async () => {
+      await expect(deleteObject(root)).to.be.rejectedWith(
         'storage/invalid-root-operation'
       );
     });
-    it('getMetadata throws', () => {
-      testShared.assertThrows(
-        root.getMetadata.bind(root),
+    it('getMetadata throws', async () => {
+      await expect(getMetadata(root)).to.be.rejectedWith(
         'storage/invalid-root-operation'
       );
     });
-    it("listAll doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        root.listAll();
-      });
-    });
-    it("list doesn't throw", () => {
-      assert.doesNotThrow(() => {
-        root.list();
-      });
-      assert.doesNotThrow(() => {
-        root.list({ pageToken: 'xxx', maxResults: 4 });
-      });
-    });
-    it('updateMetadata throws', () => {
-      testShared.assertThrows(
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        (root as any).updateMetadata.bind(root, {}),
+    it('updateMetadata throws', async () => {
+      await expect(updateMetadata(root, {} as Metadata)).to.be.rejectedWith(
         'storage/invalid-root-operation'
       );
     });
-    it('getDownloadURL throws', () => {
-      testShared.assertThrows(
-        root.getDownloadURL.bind(root),
+    it('getDownloadURL throws', async () => {
+      await expect(getDownloadURL(root)).to.be.rejectedWith(
         'storage/invalid-root-operation'
       );
     });
