@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2017 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { assert } from 'chai';
+import { expect } from 'chai';
 import { TaskEvent } from '../../src/implementation/taskenums';
 import { XhrIoPool } from '../../src/implementation/xhriopool';
-import { StorageService } from '../../src/service';
+import { StorageService, ref } from '../../src/service';
 import * as testShared from './testshared';
 import { DEFAULT_HOST } from '../../src/implementation/constants';
 import { FirebaseStorageError } from '../../src/implementation/error';
+import {
+  Reference,
+  getMetadata,
+  uploadBytesResumable
+} from '../../src/reference';
+import { Location } from '../../src/implementation/location';
 
 const fakeAppGs = testShared.makeFakeApp('gs://mybucket');
 const fakeAppGsEndingSlash = testShared.makeFakeApp('gs://mybucket/');
 const fakeAppInvalidGs = testShared.makeFakeApp('gs://mybucket/hello');
 const xhrIoPool = new XhrIoPool();
+const testLocation = new Location('bucket', 'object');
 
 function makeGsUrl(child: string = ''): string {
   return 'gs://' + testShared.bucket + '/' + child;
@@ -39,30 +46,12 @@ describe('Firebase Storage > Service', () => {
       xhrIoPool
     );
     it('Root refs point to the right place', () => {
-      const ref = service.ref();
-      assert.equal(ref.toString(), makeGsUrl());
+      const reference = ref(service);
+      expect(reference.toString()).to.equal(makeGsUrl());
     });
     it('Child refs point to the right place', () => {
-      const ref = service.ref('path/to/child');
-      assert.equal(ref.toString(), makeGsUrl('path/to/child'));
-    });
-    it('Throws calling ref with a gs:// URL', () => {
-      const error = testShared.assertThrows(() => {
-        service.ref('gs://bucket/object');
-      }, 'storage/invalid-argument');
-      assert.match(error.message, /refFromURL/);
-    });
-    it('Throws calling ref with an http:// URL', () => {
-      const error = testShared.assertThrows(() => {
-        service.ref(`http://${DEFAULT_HOST}/etc`);
-      }, 'storage/invalid-argument');
-      assert.match(error.message, /refFromURL/);
-    });
-    it('Throws calling ref with an https:// URL', () => {
-      const error = testShared.assertThrows(() => {
-        service.ref(`https://${DEFAULT_HOST}/etc`);
-      }, 'storage/invalid-argument');
-      assert.match(error.message, /refFromURL/);
+      const reference = ref(service, 'path/to/child');
+      expect(reference.toString()).to.equal(makeGsUrl('path/to/child'));
     });
   });
   describe('custom bucket constructor', () => {
@@ -73,8 +62,8 @@ describe('Firebase Storage > Service', () => {
         xhrIoPool,
         'gs://foo-bar.appspot.com'
       );
-      const ref = service.ref();
-      assert.equal(ref.toString(), 'gs://foo-bar.appspot.com/');
+      const reference = ref(service);
+      expect(reference.toString()).to.equal('gs://foo-bar.appspot.com/');
     });
     it('http:// custom bucket constructor refs point to the right place', () => {
       const service = new StorageService(
@@ -83,8 +72,8 @@ describe('Firebase Storage > Service', () => {
         xhrIoPool,
         `http://${DEFAULT_HOST}/v1/b/foo-bar.appspot.com/o`
       );
-      const ref = service.ref();
-      assert.equal(ref.toString(), 'gs://foo-bar.appspot.com/');
+      const reference = ref(service);
+      expect(reference.toString()).to.equal('gs://foo-bar.appspot.com/');
     });
     it('https:// custom bucket constructor refs point to the right place', () => {
       const service = new StorageService(
@@ -93,8 +82,8 @@ describe('Firebase Storage > Service', () => {
         xhrIoPool,
         `https://${DEFAULT_HOST}/v1/b/foo-bar.appspot.com/o`
       );
-      const ref = service.ref();
-      assert.equal(ref.toString(), 'gs://foo-bar.appspot.com/');
+      const reference = ref(service);
+      expect(reference.toString()).to.equal('gs://foo-bar.appspot.com/');
     });
 
     it('Bare bucket name constructor refs point to the right place', () => {
@@ -104,8 +93,8 @@ describe('Firebase Storage > Service', () => {
         xhrIoPool,
         'foo-bar.appspot.com'
       );
-      const ref = service.ref();
-      assert.equal(ref.toString(), 'gs://foo-bar.appspot.com/');
+      const reference = ref(service);
+      expect(reference.toString()).to.equal('gs://foo-bar.appspot.com/');
     });
     it('Child refs point to the right place', () => {
       const service = new StorageService(
@@ -114,8 +103,10 @@ describe('Firebase Storage > Service', () => {
         xhrIoPool,
         'foo-bar.appspot.com'
       );
-      const ref = service.ref('path/to/child');
-      assert.equal(ref.toString(), 'gs://foo-bar.appspot.com/path/to/child');
+      const reference = ref(service, 'path/to/child');
+      expect(reference.toString()).to.equal(
+        'gs://foo-bar.appspot.com/path/to/child'
+      );
     });
     it('Throws trying to construct with a gs:// URL containing an object path', () => {
       const error = testShared.assertThrows(() => {
@@ -126,7 +117,7 @@ describe('Firebase Storage > Service', () => {
           'gs://bucket/object/'
         );
       }, 'storage/invalid-default-bucket');
-      assert.match(error.message, /Invalid default bucket/);
+      expect(error.message).to.match(/Invalid default bucket/);
     });
   });
   describe('default bucket config', () => {
@@ -136,7 +127,7 @@ describe('Firebase Storage > Service', () => {
         testShared.fakeAuthProvider,
         xhrIoPool
       );
-      assert.equal(service.ref().toString(), 'gs://mybucket/');
+      expect(ref(service)?.toString()).to.equal('gs://mybucket/');
     });
     it('gs:// works with ending slash', () => {
       const service = new StorageService(
@@ -144,7 +135,7 @@ describe('Firebase Storage > Service', () => {
         testShared.fakeAuthProvider,
         xhrIoPool
       );
-      assert.equal(service.ref().toString(), 'gs://mybucket/');
+      expect(ref(service)?.toString()).to.equal('gs://mybucket/');
     });
     it('Throws when config bucket is gs:// with an object path', () => {
       testShared.assertThrows(() => {
@@ -156,102 +147,131 @@ describe('Firebase Storage > Service', () => {
       }, 'storage/invalid-default-bucket');
     });
   });
-  describe('refFromURL', () => {
+  describe('ref(service, url)', () => {
     const service = new StorageService(
       testShared.fakeApp,
       testShared.fakeAuthProvider,
       xhrIoPool
     );
-    it('Throws on non-URL arg', () => {
-      const error = testShared.assertThrows(() => {
-        service.refFromURL('path/to/child');
-      }, 'storage/invalid-argument');
-      assert.match(
-        error.message,
-        /Expected full URL but got a child path, use ref instead/i
+    it('Works with gs:// URLs', () => {
+      const reference = ref(service, 'gs://mybucket/child/path/image.png');
+      expect(reference.toString()).to.equal(
+        'gs://mybucket/child/path/image.png'
       );
     });
-    it('Works with gs:// URLs', () => {
-      const ref = service.refFromURL('gs://mybucket/child/path/image.png');
-      assert.equal(ref.toString(), 'gs://mybucket/child/path/image.png');
-    });
     it('Works with http:// URLs', () => {
-      const ref = service.refFromURL(
+      const reference = ref(
+        service,
         `http://${DEFAULT_HOST}/v0/b/` +
           'mybucket/o/child%2Fpath%2Fimage.png?downloadToken=hello'
       );
-      assert.equal(ref.toString(), 'gs://mybucket/child/path/image.png');
+      expect(reference.toString()).to.equal(
+        'gs://mybucket/child/path/image.png'
+      );
     });
     it('Works with https:// URLs', () => {
-      const ref = service.refFromURL(
+      const reference = ref(
+        service,
         `https://${DEFAULT_HOST}/v0/b/` +
           'mybucket/o/child%2Fpath%2Fimage.png?downloadToken=hello'
       );
-      assert.equal(ref.toString(), 'gs://mybucket/child/path/image.png');
+      expect(reference.toString()).to.equal(
+        'gs://mybucket/child/path/image.png'
+      );
     });
     it('Works with storage.googleapis.com URLs', () => {
-      const ref = service.refFromURL(
+      const reference = ref(
+        service,
         `https://storage.googleapis.com/mybucket/path%20with%20space/image.png`
       );
-      assert.equal(ref.toString(), 'gs://mybucket/path with space/image.png');
+      expect(reference.toString()).to.equal(
+        'gs://mybucket/path with space/image.png'
+      );
     });
     it('Works with storage.googleapis.com URLs with query params', () => {
-      const ref = service.refFromURL(
+      const reference = ref(
+        service,
         `https://storage.googleapis.com/mybucket/path%20with%20space/image.png?X-Goog-Algorithm=
 GOOG4-RSA-SHA256`
       );
-      assert.equal(ref.toString(), 'gs://mybucket/path with space/image.png');
+      expect(reference.toString()).to.equal(
+        'gs://mybucket/path with space/image.png'
+      );
     });
     it('Works with storage.cloud.google.com URLs', () => {
-      const ref = service.refFromURL(
+      const reference = ref(
+        service,
         `https://storage.cloud.google.com/mybucket/path%20with%20space/image.png`
       );
-      assert.equal(ref.toString(), 'gs://mybucket/path with space/image.png');
+      expect(reference.toString()).to.equal(
+        'gs://mybucket/path with space/image.png'
+      );
     });
     it('Works with storage.cloud.google.com URLs and escaped slash', () => {
-      const ref = service.refFromURL(
+      const reference = ref(
+        service,
         `https://storage.cloud.google.com/mybucket/path%20with%20space%2Fimage.png`
       );
-      assert.equal(ref.toString(), 'gs://mybucket/path with space/image.png');
+      expect(reference.toString()).to.equal(
+        'gs://mybucket/path with space/image.png'
+      );
     });
   });
-  describe('Argument verification', () => {
+  describe('ref(service, path)', () => {
     const service = new StorageService(
       testShared.fakeApp,
       testShared.fakeAuthProvider,
       xhrIoPool
     );
-    describe('ref', () => {
-      it('Throws on gs:// argument', () => {
-        testShared.assertThrows(
-          testShared.bind(service.ref, service, 'gs://yo'),
-          'storage/invalid-argument'
-        );
-      });
+    it('Works with non URL paths', () => {
+      const newRef = ref(service, 'child/path/image.png');
+      expect(newRef.toString()).to.equal('gs://mybucket/child/path/image.png');
     });
-    describe('refFromURL', () => {
-      it('Throws with an invalid URL arg', () => {
-        testShared.assertThrows(
-          testShared.bind(service.refFromURL, service, 'notlegit://url'),
-          'storage/invalid-argument'
-        );
-      });
+    it('Works with no path', () => {
+      const newRef = ref(service);
+      expect(newRef.toString()).to.equal('gs://mybucket/');
     });
-    describe('setMaxUploadRetryTime', () => {
-      it('Throws on negative arg', () => {
-        testShared.assertThrows(
-          testShared.bind(service.setMaxUploadRetryTime, service, -10),
-          'storage/invalid-argument'
-        );
-      });
+  });
+  describe('ref(reference, path)', () => {
+    const service = new StorageService(
+      testShared.fakeApp,
+      testShared.fakeAuthProvider,
+      xhrIoPool
+    );
+    const reference = new Reference(service, testLocation);
+    it('Throws calling ref(reference, path) with a gs:// URL', () => {
+      const error = testShared.assertThrows(() => {
+        ref(reference, 'gs://bucket/object');
+      }, 'storage/invalid-argument');
+      expect(error.message).to.match(/url/);
     });
-    describe('setMaxOperationRetryTime', () => {
-      it('Throws on negative arg', () => {
-        testShared.assertThrows(
-          testShared.bind(service.setMaxOperationRetryTime, service, -10),
-          'storage/invalid-argument'
-        );
-      });
+    it('Throws calling ref(reference, path) with an http:// URL', () => {
+      const error = testShared.assertThrows(() => {
+        ref(reference, `http://${DEFAULT_HOST}/etc`);
+      }, 'storage/invalid-argument');
+      expect(error.message).to.match(/url/);
+    });
+    it('Throws calling ref(reference, path) with an https:// URL', () => {
+      const error = testShared.assertThrows(() => {
+        ref(reference, `https://${DEFAULT_HOST}/etc`);
+      }, 'storage/invalid-argument');
+      expect(error.message).to.match(/url/);
+    });
+    it('Works with non URL paths', () => {
+      const newRef = ref(reference, 'child/path/image.png');
+      expect(newRef.toString()).to.equal(
+        'gs://bucket/object/child/path/image.png'
+      );
+    });
+    it('Works with no path', () => {
+      const newRef = ref(reference);
+      expect(newRef.toString()).to.equal('gs://bucket/object');
+    });
+    it('Throws calling ref(reference, path) if path contains ".."', () => {
+      const error = testShared.assertThrows(() => {
+        ref(reference, `../child/path`);
+      }, 'storage/invalid-argument');
+      expect(error.message).to.match(/"\.\."/);
     });
   });
 
@@ -261,53 +281,39 @@ GOOG4-RSA-SHA256`
       testShared.fakeAuthProvider,
       xhrIoPool
     );
-    it('In-flight requests are canceled when the service is deleted', () => {
-      const ref = service.refFromURL('gs://mybucket/image.jpg');
-      const toReturn = ref.getMetadata().then(
-        () => {
-          assert.fail('Promise succeeded, should have been canceled');
-        },
-        err => {
-          assert.equal(err.code, 'storage/app-deleted');
-        }
-      );
+    it('In-flight requests are canceled when the service is deleted', async () => {
+      const reference = ref(service, 'gs://mybucket/image.jpg');
+      const metadataPromise = getMetadata(reference);
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      service.INTERNAL.delete();
-      return toReturn;
+      service._delete();
+      await expect(metadataPromise).to.be.rejectedWith('storage/app-deleted');
     });
-    it('Requests fail when started after the service is deleted', () => {
-      const ref = service.refFromURL('gs://mybucket/image.jpg');
+    it('Requests fail when started after the service is deleted', async () => {
+      const reference = ref(service, 'gs://mybucket/image.jpg');
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      service.INTERNAL.delete();
-      const toReturn = ref.getMetadata().then(
-        () => {
-          assert.fail('Promise succeeded, should have been canceled');
-        },
-        err => {
-          assert.equal(err.code, 'storage/app-deleted');
-        }
+      service._delete();
+      await expect(getMetadata(reference)).to.be.rejectedWith(
+        'storage/app-deleted'
       );
-      return toReturn;
     });
     it('Running uploads fail when the service is deleted', () => {
-      const ref = service.refFromURL('gs://mybucket/image.jpg');
+      const reference = ref(service, 'gs://mybucket/image.jpg');
       const toReturn = new Promise((resolve, reject) => {
-        ref.put(new Blob(['a'])).on(
+        uploadBytesResumable(reference, new Blob(['a'])).on(
           TaskEvent.STATE_CHANGED,
-          null,
+          undefined,
           (err: FirebaseStorageError | Error) => {
-            assert.equal(
-              (err as FirebaseStorageError).code,
+            expect((err as FirebaseStorageError).code).to.equal(
               'storage/app-deleted'
             );
             resolve();
           },
           () => {
-            assert.fail('Upload completed, should have been canceled');
+            reject('Upload completed, should have been canceled');
           }
         );
         // eslint-disable-next-line @typescript-eslint/no-floating-promises
-        service.INTERNAL.delete();
+        service._delete();
       });
       return toReturn;
     });
