@@ -80,7 +80,6 @@ import {
 } from '../../../src/core/event_manager';
 import { FirestoreError } from '../../../src/util/error';
 import { Compat } from '../../../src/compat/compat';
-import { _BaseFieldPath } from '../../../src/api/field_path';
 
 /**
  * An options object that can be passed to {@link onSnapshot()} and {@link
@@ -384,7 +383,7 @@ export function updateDoc(
   let parsed: ParsedUpdateData;
   if (
     typeof fieldOrUpdateData === 'string' ||
-    fieldOrUpdateData instanceof _BaseFieldPath
+    fieldOrUpdateData instanceof FieldPath
   ) {
     parsed = parseUpdateVarargs(
       dataReader,
