@@ -25,7 +25,7 @@ import { StringFormat, dataFromString } from './string';
 import * as type from './type';
 
 /**
- * @param opt_elideCopy If true, doesn't copy mutable input data
+ * @param opt_elideCopy - If true, doesn't copy mutable input data
  *     (e.g. Uint8Arrays). Pass true only if you know the objects will not be
  *     modified after this blob's construction.
  */
