@@ -18,52 +18,47 @@ import * as type from './type';
 import { FirebaseStorageError } from './error';
 
 export type NextFn<T> = (value: T) => void;
-export type ErrorFn = (error: Error | FirebaseStorageError) => void;
+export type ErrorFn = (error: FirebaseStorageError) => void;
 export type CompleteFn = () => void;
 export type Unsubscribe = () => void;
 
 export interface StorageObserver<T> {
-  next?: NextFn<T> | null;
-  error?: ErrorFn | null;
-  complete?: CompleteFn | null;
+  next?: NextFn<T>;
+  error?: ErrorFn;
+  complete?: CompleteFn;
 }
 
 export type Subscribe<T> = (
-  next?: NextFn<T> | StorageObserver<T> | null,
-  error?: ErrorFn | null,
-  complete?: CompleteFn | null
+  next?: NextFn<T> | StorageObserver<T>,
+  error?: ErrorFn,
+  complete?: CompleteFn
 ) => Unsubscribe;
 
-/**
- * @struct
- */
 export class Observer<T> implements StorageObserver<T> {
-  next?: NextFn<T> | null;
-  error?: ErrorFn | null;
-  complete?: CompleteFn | null;
+  next?: NextFn<T>;
+  error?: ErrorFn;
+  complete?: CompleteFn;
 
   constructor(
-    nextOrObserver?: NextFn<T> | StorageObserver<T> | null,
-    error?: ErrorFn | null,
-    complete?: CompleteFn | null
+    nextOrObserver?: NextFn<T> | StorageObserver<T>,
+    error?: ErrorFn,
+    complete?: CompleteFn
   ) {
     const asFunctions =
-      type.isFunction(nextOrObserver) ||
-      type.isDef(error) ||
-      type.isDef(complete);
+      type.isFunction(nextOrObserver) || error != null || complete != null;
     if (asFunctions) {
-      this.next = nextOrObserver as NextFn<T> | null;
-      this.error = error || null;
-      this.complete = complete || null;
+      this.next = nextOrObserver as NextFn<T>;
+      this.error = error;
+      this.complete = complete;
     } else {
       const observer = nextOrObserver as {
-        next?: NextFn<T> | null;
-        error?: ErrorFn | null;
-        complete?: CompleteFn | null;
+        next?: NextFn<T>;
+        error?: ErrorFn;
+        complete?: CompleteFn;
       };
-      this.next = observer.next || null;
-      this.error = observer.error || null;
-      this.complete = observer.complete || null;
+      this.next = observer.next;
+      this.error = observer.error;
+      this.complete = observer.complete;
     }
   }
 }
