@@ -30,7 +30,6 @@ import { Metadata } from './metadata';
 import {
   CompleteFn,
   ErrorFn,
-  NextFn,
   Observer,
   StorageObserver,
   Subscribe,
@@ -39,140 +38,140 @@ import {
 import { Request } from './implementation/request';
 import { UploadTaskSnapshot } from './tasksnapshot';
 import { async as fbsAsync } from './implementation/async';
-import { Location } from './implementation/location';
 import * as fbsMetadata from './implementation/metadata';
 import * as fbsRequests from './implementation/requests';
 import { Reference } from './reference';
-import { StorageService } from './service';
+import { getMappings } from './implementation/metadata';
 
 /**
  * Represents a blob being uploaded. Can be used to pause/resume/cancel the
  * upload and manage callbacks for various events.
  */
 export class UploadTask {
-  private ref_: Reference;
-  private service_: StorageService;
-  private location_: Location;
-  private blob_: FbsBlob;
-  private metadata_: Metadata | null;
-  private mappings_: fbsMetadata.Mappings;
-  private transferred_: number = 0;
-  private needToFetchStatus_: boolean = false;
-  private needToFetchMetadata_: boolean = false;
-  private observers_: Array<Observer<UploadTaskSnapshot>> = [];
-  private resumable_: boolean;
-  private state_: InternalTaskState;
-  private error_: Error | null = null;
-  private uploadUrl_: string | null = null;
-  private request_: Request<unknown> | null = null;
-  private chunkMultiplier_: number = 1;
-  private errorHandler_: (p1: FirebaseStorageError) => void;
-  private metadataErrorHandler_: (p1: FirebaseStorageError) => void;
-  private resolve_: ((p1: UploadTaskSnapshot) => void) | null = null;
-  private reject_: ((p1: Error) => void) | null = null;
-  private promise_: Promise<UploadTaskSnapshot>;
+  private _ref: Reference;
+  /**
+   * @internal
+   */
+  _blob: FbsBlob;
+  /**
+   * @internal
+   */
+  _metadata: Metadata | null;
+  private _mappings: fbsMetadata.Mappings;
+  /**
+   * @internal
+   */
+  _transferred: number = 0;
+  private _needToFetchStatus: boolean = false;
+  private _needToFetchMetadata: boolean = false;
+  private _observers: Array<StorageObserver<UploadTaskSnapshot>> = [];
+  private _resumable: boolean;
+  /**
+   * @internal
+   */
+  _state: InternalTaskState;
+  private _error?: FirebaseStorageError = undefined;
+  private _uploadUrl?: string = undefined;
+  private _request?: Request<unknown> = undefined;
+  private _chunkMultiplier: number = 1;
+  private _errorHandler: (p1: FirebaseStorageError) => void;
+  private _metadataErrorHandler: (p1: FirebaseStorageError) => void;
+  private _resolve?: (p1: UploadTaskSnapshot) => void = undefined;
+  private _reject?: (p1: FirebaseStorageError) => void = undefined;
+  private _promise: Promise<UploadTaskSnapshot>;
 
   /**
-   * @param ref The firebaseStorage.Reference object this task came
+   * @param ref - The firebaseStorage.Reference object this task came
    *     from, untyped to avoid cyclic dependencies.
-   * @param blob The blob to upload.
+   * @param blob - The blob to upload.
    */
-  constructor(
-    ref: Reference,
-    service: StorageService,
-    location: Location,
-    mappings: fbsMetadata.Mappings,
-    blob: FbsBlob,
-    metadata: Metadata | null = null
-  ) {
-    this.ref_ = ref;
-    this.service_ = service;
-    this.location_ = location;
-    this.blob_ = blob;
-    this.metadata_ = metadata;
-    this.mappings_ = mappings;
-    this.resumable_ = this.shouldDoResumable_(this.blob_);
-    this.state_ = InternalTaskState.RUNNING;
-    this.errorHandler_ = error => {
-      this.request_ = null;
-      this.chunkMultiplier_ = 1;
+  constructor(ref: Reference, blob: FbsBlob, metadata: Metadata | null = null) {
+    this._ref = ref;
+    this._blob = blob;
+    this._metadata = metadata;
+    this._mappings = getMappings();
+    this._resumable = this._shouldDoResumable(this._blob);
+    this._state = InternalTaskState.RUNNING;
+    this._errorHandler = error => {
+      this._request = undefined;
+      this._chunkMultiplier = 1;
       if (error.codeEquals(Code.CANCELED)) {
-        this.needToFetchStatus_ = true;
+        this._needToFetchStatus = true;
         this.completeTransitions_();
       } else {
-        this.error_ = error;
-        this.transition_(InternalTaskState.ERROR);
+        this._error = error;
+        this._transition(InternalTaskState.ERROR);
       }
     };
-    this.metadataErrorHandler_ = error => {
-      this.request_ = null;
+    this._metadataErrorHandler = error => {
+      this._request = undefined;
       if (error.codeEquals(Code.CANCELED)) {
         this.completeTransitions_();
       } else {
-        this.error_ = error;
-        this.transition_(InternalTaskState.ERROR);
+        this._error = error;
+        this._transition(InternalTaskState.ERROR);
       }
     };
-    this.promise_ = new Promise((resolve, reject) => {
-      this.resolve_ = resolve;
-      this.reject_ = reject;
-      this.start_();
+    this._promise = new Promise((resolve, reject) => {
+      this._resolve = resolve;
+      this._reject = reject;
+      this._start();
     });
 
     // Prevent uncaught rejections on the internal promise from bubbling out
     // to the top level with a dummy handler.
-    this.promise_.then(null, () => {});
+    this._promise.then(null, () => {});
   }
 
-  private makeProgressCallback_(): (p1: number, p2: number) => void {
-    const sizeBefore = this.transferred_;
-    return loaded => this.updateProgress_(sizeBefore + loaded);
+  private _makeProgressCallback(): (p1: number, p2: number) => void {
+    const sizeBefore = this._transferred;
+    return loaded => this._updateProgress(sizeBefore + loaded);
   }
 
-  private shouldDoResumable_(blob: FbsBlob): boolean {
+  private _shouldDoResumable(blob: FbsBlob): boolean {
     return blob.size() > 256 * 1024;
   }
 
-  private start_(): void {
-    if (this.state_ !== InternalTaskState.RUNNING) {
+  private _start(): void {
+    if (this._state !== InternalTaskState.RUNNING) {
       // This can happen if someone pauses us in a resume callback, for example.
       return;
     }
-    if (this.request_ !== null) {
+    if (this._request !== undefined) {
       return;
     }
-    if (this.resumable_) {
-      if (this.uploadUrl_ === null) {
-        this.createResumable_();
+    if (this._resumable) {
+      if (this._uploadUrl === undefined) {
+        this._createResumable();
       } else {
-        if (this.needToFetchStatus_) {
-          this.fetchStatus_();
+        if (this._needToFetchStatus) {
+          this._fetchStatus();
         } else {
-          if (this.needToFetchMetadata_) {
+          if (this._needToFetchMetadata) {
             // Happens if we miss the metadata on upload completion.
-            this.fetchMetadata_();
+            this._fetchMetadata();
           } else {
-            this.continueUpload_();
+            this._continueUpload();
           }
         }
       }
     } else {
-      this.oneShotUpload_();
+      this._oneShotUpload();
     }
   }
 
-  private resolveToken_(callback: (p1: string | null) => void): void {
+  private _resolveToken(callback: (p1: string | null) => void): void {
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    this.service_.getAuthToken().then(authToken => {
-      switch (this.state_) {
+    this._ref.storage.getAuthToken().then(authToken => {
+      switch (this._state) {
         case InternalTaskState.RUNNING:
           callback(authToken);
           break;
         case InternalTaskState.CANCELING:
-          this.transition_(InternalTaskState.CANCELED);
+          this._transition(InternalTaskState.CANCELED);
           break;
         case InternalTaskState.PAUSING:
-          this.transition_(InternalTaskState.PAUSED);
+          this._transition(InternalTaskState.PAUSED);
           break;
         default:
       }
@@ -181,161 +180,173 @@ export class UploadTask {
 
   // TODO(andysoto): assert false
 
-  private createResumable_(): void {
-    this.resolveToken_(authToken => {
+  private _createResumable(): void {
+    this._resolveToken(authToken => {
       const requestInfo = fbsRequests.createResumableUpload(
-        this.service_,
-        this.location_,
-        this.mappings_,
-        this.blob_,
-        this.metadata_
+        this._ref.storage,
+        this._ref._location,
+        this._mappings,
+        this._blob,
+        this._metadata
       );
-      const createRequest = this.service_.makeRequest(requestInfo, authToken);
-      this.request_ = createRequest;
-      createRequest.getPromise().then((url: string) => {
-        this.request_ = null;
-        this.uploadUrl_ = url;
-        this.needToFetchStatus_ = false;
-        this.completeTransitions_();
-      }, this.errorHandler_);
-    });
-  }
-
-  private fetchStatus_(): void {
-    // TODO(andysoto): assert(this.uploadUrl_ !== null);
-    const url = this.uploadUrl_ as string;
-    this.resolveToken_(authToken => {
-      const requestInfo = fbsRequests.getResumableUploadStatus(
-        this.service_,
-        this.location_,
-        url,
-        this.blob_
-      );
-      const statusRequest = this.service_.makeRequest(requestInfo, authToken);
-      this.request_ = statusRequest;
-      statusRequest.getPromise().then(status => {
-        status = status as fbsRequests.ResumableUploadStatus;
-        this.request_ = null;
-        this.updateProgress_(status.current);
-        this.needToFetchStatus_ = false;
-        if (status.finalized) {
-          this.needToFetchMetadata_ = true;
-        }
-        this.completeTransitions_();
-      }, this.errorHandler_);
-    });
-  }
-
-  private continueUpload_(): void {
-    const chunkSize =
-      fbsRequests.resumableUploadChunkSize * this.chunkMultiplier_;
-    const status = new fbsRequests.ResumableUploadStatus(
-      this.transferred_,
-      this.blob_.size()
-    );
-
-    // TODO(andysoto): assert(this.uploadUrl_ !== null);
-    const url = this.uploadUrl_ as string;
-    this.resolveToken_(authToken => {
-      let requestInfo;
-      try {
-        requestInfo = fbsRequests.continueResumableUpload(
-          this.location_,
-          this.service_,
-          url,
-          this.blob_,
-          chunkSize,
-          this.mappings_,
-          status,
-          this.makeProgressCallback_()
-        );
-      } catch (e) {
-        this.error_ = e;
-        this.transition_(InternalTaskState.ERROR);
-        return;
-      }
-      const uploadRequest = this.service_.makeRequest(requestInfo, authToken);
-      this.request_ = uploadRequest;
-      uploadRequest
-        .getPromise()
-        .then((newStatus: fbsRequests.ResumableUploadStatus) => {
-          this.increaseMultiplier_();
-          this.request_ = null;
-          this.updateProgress_(newStatus.current);
-          if (newStatus.finalized) {
-            this.metadata_ = newStatus.metadata;
-            this.transition_(InternalTaskState.SUCCESS);
-          } else {
-            this.completeTransitions_();
-          }
-        }, this.errorHandler_);
-    });
-  }
-
-  private increaseMultiplier_(): void {
-    const currentSize =
-      fbsRequests.resumableUploadChunkSize * this.chunkMultiplier_;
-
-    // Max chunk size is 32M.
-    if (currentSize < 32 * 1024 * 1024) {
-      this.chunkMultiplier_ *= 2;
-    }
-  }
-
-  private fetchMetadata_(): void {
-    this.resolveToken_(authToken => {
-      const requestInfo = fbsRequests.getMetadata(
-        this.service_,
-        this.location_,
-        this.mappings_
-      );
-      const metadataRequest = this.service_.makeRequest(requestInfo, authToken);
-      this.request_ = metadataRequest;
-      metadataRequest.getPromise().then(metadata => {
-        this.request_ = null;
-        this.metadata_ = metadata;
-        this.transition_(InternalTaskState.SUCCESS);
-      }, this.metadataErrorHandler_);
-    });
-  }
-
-  private oneShotUpload_(): void {
-    this.resolveToken_(authToken => {
-      const requestInfo = fbsRequests.multipartUpload(
-        this.service_,
-        this.location_,
-        this.mappings_,
-        this.blob_,
-        this.metadata_
-      );
-      const multipartRequest = this.service_.makeRequest(
+      const createRequest = this._ref.storage.makeRequest(
         requestInfo,
         authToken
       );
-      this.request_ = multipartRequest;
-      multipartRequest.getPromise().then(metadata => {
-        this.request_ = null;
-        this.metadata_ = metadata;
-        this.updateProgress_(this.blob_.size());
-        this.transition_(InternalTaskState.SUCCESS);
-      }, this.errorHandler_);
+      this._request = createRequest;
+      createRequest.getPromise().then((url: string) => {
+        this._request = undefined;
+        this._uploadUrl = url;
+        this._needToFetchStatus = false;
+        this.completeTransitions_();
+      }, this._errorHandler);
     });
   }
 
-  private updateProgress_(transferred: number): void {
-    const old = this.transferred_;
-    this.transferred_ = transferred;
+  private _fetchStatus(): void {
+    // TODO(andysoto): assert(this.uploadUrl_ !== null);
+    const url = this._uploadUrl as string;
+    this._resolveToken(authToken => {
+      const requestInfo = fbsRequests.getResumableUploadStatus(
+        this._ref.storage,
+        this._ref._location,
+        url,
+        this._blob
+      );
+      const statusRequest = this._ref.storage.makeRequest(
+        requestInfo,
+        authToken
+      );
+      this._request = statusRequest;
+      statusRequest.getPromise().then(status => {
+        status = status as fbsRequests.ResumableUploadStatus;
+        this._request = undefined;
+        this._updateProgress(status.current);
+        this._needToFetchStatus = false;
+        if (status.finalized) {
+          this._needToFetchMetadata = true;
+        }
+        this.completeTransitions_();
+      }, this._errorHandler);
+    });
+  }
+
+  private _continueUpload(): void {
+    const chunkSize =
+      fbsRequests.resumableUploadChunkSize * this._chunkMultiplier;
+    const status = new fbsRequests.ResumableUploadStatus(
+      this._transferred,
+      this._blob.size()
+    );
+
+    // TODO(andysoto): assert(this.uploadUrl_ !== null);
+    const url = this._uploadUrl as string;
+    this._resolveToken(authToken => {
+      let requestInfo;
+      try {
+        requestInfo = fbsRequests.continueResumableUpload(
+          this._ref._location,
+          this._ref.storage,
+          url,
+          this._blob,
+          chunkSize,
+          this._mappings,
+          status,
+          this._makeProgressCallback()
+        );
+      } catch (e) {
+        this._error = e;
+        this._transition(InternalTaskState.ERROR);
+        return;
+      }
+      const uploadRequest = this._ref.storage.makeRequest(
+        requestInfo,
+        authToken
+      );
+      this._request = uploadRequest;
+      uploadRequest
+        .getPromise()
+        .then((newStatus: fbsRequests.ResumableUploadStatus) => {
+          this._increaseMultiplier();
+          this._request = undefined;
+          this._updateProgress(newStatus.current);
+          if (newStatus.finalized) {
+            this._metadata = newStatus.metadata;
+            this._transition(InternalTaskState.SUCCESS);
+          } else {
+            this.completeTransitions_();
+          }
+        }, this._errorHandler);
+    });
+  }
+
+  private _increaseMultiplier(): void {
+    const currentSize =
+      fbsRequests.resumableUploadChunkSize * this._chunkMultiplier;
+
+    // Max chunk size is 32M.
+    if (currentSize < 32 * 1024 * 1024) {
+      this._chunkMultiplier *= 2;
+    }
+  }
+
+  private _fetchMetadata(): void {
+    this._resolveToken(authToken => {
+      const requestInfo = fbsRequests.getMetadata(
+        this._ref.storage,
+        this._ref._location,
+        this._mappings
+      );
+      const metadataRequest = this._ref.storage.makeRequest(
+        requestInfo,
+        authToken
+      );
+      this._request = metadataRequest;
+      metadataRequest.getPromise().then(metadata => {
+        this._request = undefined;
+        this._metadata = metadata;
+        this._transition(InternalTaskState.SUCCESS);
+      }, this._metadataErrorHandler);
+    });
+  }
+
+  private _oneShotUpload(): void {
+    this._resolveToken(authToken => {
+      const requestInfo = fbsRequests.multipartUpload(
+        this._ref.storage,
+        this._ref._location,
+        this._mappings,
+        this._blob,
+        this._metadata
+      );
+      const multipartRequest = this._ref.storage.makeRequest(
+        requestInfo,
+        authToken
+      );
+      this._request = multipartRequest;
+      multipartRequest.getPromise().then(metadata => {
+        this._request = undefined;
+        this._metadata = metadata;
+        this._updateProgress(this._blob.size());
+        this._transition(InternalTaskState.SUCCESS);
+      }, this._errorHandler);
+    });
+  }
+
+  private _updateProgress(transferred: number): void {
+    const old = this._transferred;
+    this._transferred = transferred;
 
     // A progress update can make the "transferred" value smaller (e.g. a
     // partial upload not completed by server, after which the "transferred"
     // value may reset to the value at the beginning of the request).
-    if (this.transferred_ !== old) {
-      this.notifyObservers_();
+    if (this._transferred !== old) {
+      this._notifyObservers();
     }
   }
 
-  private transition_(state: InternalTaskState): void {
-    if (this.state_ === state) {
+  private _transition(state: InternalTaskState): void {
+    if (this._state === state) {
       return;
     }
     switch (state) {
@@ -343,74 +354,74 @@ export class UploadTask {
         // TODO(andysoto):
         // assert(this.state_ === InternalTaskState.RUNNING ||
         //        this.state_ === InternalTaskState.PAUSING);
-        this.state_ = state;
-        if (this.request_ !== null) {
-          this.request_.cancel();
+        this._state = state;
+        if (this._request !== undefined) {
+          this._request.cancel();
         }
         break;
       case InternalTaskState.PAUSING:
         // TODO(andysoto):
         // assert(this.state_ === InternalTaskState.RUNNING);
-        this.state_ = state;
-        if (this.request_ !== null) {
-          this.request_.cancel();
+        this._state = state;
+        if (this._request !== undefined) {
+          this._request.cancel();
         }
         break;
       case InternalTaskState.RUNNING:
         // TODO(andysoto):
         // assert(this.state_ === InternalTaskState.PAUSED ||
         //        this.state_ === InternalTaskState.PAUSING);
-        const wasPaused = this.state_ === InternalTaskState.PAUSED;
-        this.state_ = state;
+        const wasPaused = this._state === InternalTaskState.PAUSED;
+        this._state = state;
         if (wasPaused) {
-          this.notifyObservers_();
-          this.start_();
+          this._notifyObservers();
+          this._start();
         }
         break;
       case InternalTaskState.PAUSED:
         // TODO(andysoto):
         // assert(this.state_ === InternalTaskState.PAUSING);
-        this.state_ = state;
-        this.notifyObservers_();
+        this._state = state;
+        this._notifyObservers();
         break;
       case InternalTaskState.CANCELED:
         // TODO(andysoto):
         // assert(this.state_ === InternalTaskState.PAUSED ||
         //        this.state_ === InternalTaskState.CANCELING);
-        this.error_ = canceled();
-        this.state_ = state;
-        this.notifyObservers_();
+        this._error = canceled();
+        this._state = state;
+        this._notifyObservers();
         break;
       case InternalTaskState.ERROR:
         // TODO(andysoto):
         // assert(this.state_ === InternalTaskState.RUNNING ||
         //        this.state_ === InternalTaskState.PAUSING ||
         //        this.state_ === InternalTaskState.CANCELING);
-        this.state_ = state;
-        this.notifyObservers_();
+        this._state = state;
+        this._notifyObservers();
         break;
       case InternalTaskState.SUCCESS:
         // TODO(andysoto):
         // assert(this.state_ === InternalTaskState.RUNNING ||
         //        this.state_ === InternalTaskState.PAUSING ||
         //        this.state_ === InternalTaskState.CANCELING);
-        this.state_ = state;
-        this.notifyObservers_();
+        this._state = state;
+        this._notifyObservers();
         break;
       default: // Ignore
     }
   }
 
   private completeTransitions_(): void {
-    switch (this.state_) {
+    switch (this._state) {
       case InternalTaskState.PAUSING:
-        this.transition_(InternalTaskState.PAUSED);
+        this._transition(InternalTaskState.PAUSED);
         break;
       case InternalTaskState.CANCELING:
-        this.transition_(InternalTaskState.CANCELED);
+        this._transition(InternalTaskState.CANCELED);
         break;
       case InternalTaskState.RUNNING:
-        this.start_();
+        this._start();
         break;
       default:
         // TODO(andysoto): assert(false);
@@ -419,50 +430,49 @@ export class UploadTask {
   }
 
   get snapshot(): UploadTaskSnapshot {
-    const externalState = taskStateFromInternalTaskState(this.state_);
+    const externalState = taskStateFromInternalTaskState(this._state);
     return new UploadTaskSnapshot(
-      this.transferred_,
-      this.blob_.size(),
+      this._transferred,
+      this._blob.size(),
       externalState,
-      this.metadata_,
+      this._metadata!,
       this,
-      this.ref_
+      this._ref
     );
   }
 
   /**
    * Adds a callback for an event.
-   * @param type The type of event to listen for.
+   * @param type - The type of event to listen for.
    */
   on(
     type: TaskEvent,
     nextOrObserver?:
-      | NextFn<UploadTaskSnapshot>
       | StorageObserver<UploadTaskSnapshot>
-      | null,
-    error?: ErrorFn | null,
-    completed?: CompleteFn | null
+      | ((a: UploadTaskSnapshot) => unknown),
+    error?: ErrorFn,
+    completed?: CompleteFn
   ): Unsubscribe | Subscribe<UploadTaskSnapshot> {
     const observer = new Observer(nextOrObserver, error, completed);
-    this.addObserver_(observer);
+    this._addObserver(observer);
     return () => {
-      this.removeObserver_(observer);
+      this._removeObserver(observer);
     };
   }
 
   /**
    * This object behaves like a Promise, and resolves with its snapshot data
    * when the upload completes.
-   * @param onFulfilled The fulfillment callback. Promise chaining works as normal.
-   * @param onRejected The rejection callback.
+   * @param onFulfilled - The fulfillment callback. Promise chaining works as normal.
+   * @param onRejected - The rejection callback.
    */
   then<U>(
     onFulfilled?: ((value: UploadTaskSnapshot) => U | Promise<U>) | null,
-    onRejected?: ((error: Error) => U | Promise<U>) | null
+    onRejected?: ((error: FirebaseStorageError) => U | Promise<U>) | null
   ): Promise<U> {
     // These casts are needed so that TypeScript can infer the types of the
     // resulting Promise.
-    return this.promise_.then<U>(
+    return this._promise.then<U>(
       onFulfilled as (value: UploadTaskSnapshot) => U | Promise<U>,
       onRejected as ((error: unknown) => Promise<never>) | null
     );
@@ -471,61 +481,63 @@ export class UploadTask {
   /**
    * Equivalent to calling `then(null, onRejected)`.
    */
-  catch<T>(onRejected: (p1: Error) => T | Promise<T>): Promise<T> {
+  catch<T>(
+    onRejected: (p1: FirebaseStorageError) => T | Promise<T>
+  ): Promise<T> {
     return this.then(null, onRejected);
   }
 
   /**
    * Adds the given observer.
    */
-  private addObserver_(observer: Observer<UploadTaskSnapshot>): void {
-    this.observers_.push(observer);
-    this.notifyObserver_(observer);
+  private _addObserver(observer: Observer<UploadTaskSnapshot>): void {
+    this._observers.push(observer);
+    this._notifyObserver(observer);
   }
 
   /**
    * Removes the given observer.
    */
-  private removeObserver_(observer: Observer<UploadTaskSnapshot>): void {
-    const i = this.observers_.indexOf(observer);
+  private _removeObserver(observer: Observer<UploadTaskSnapshot>): void {
+    const i = this._observers.indexOf(observer);
     if (i !== -1) {
-      this.observers_.splice(i, 1);
+      this._observers.splice(i, 1);
     }
   }
 
-  private notifyObservers_(): void {
-    this.finishPromise_();
-    const observers = this.observers_.slice();
+  private _notifyObservers(): void {
+    this._finishPromise();
+    const observers = this._observers.slice();
     observers.forEach(observer => {
-      this.notifyObserver_(observer);
+      this._notifyObserver(observer);
     });
   }
 
-  private finishPromise_(): void {
-    if (this.resolve_ !== null) {
+  private _finishPromise(): void {
+    if (this._resolve !== undefined) {
       let triggered = true;
-      switch (taskStateFromInternalTaskState(this.state_)) {
+      switch (taskStateFromInternalTaskState(this._state)) {
         case TaskState.SUCCESS:
-          fbsAsync(this.resolve_.bind(null, this.snapshot))();
+          fbsAsync(this._resolve.bind(null, this.snapshot))();
           break;
         case TaskState.CANCELED:
         case TaskState.ERROR:
-          const toCall = this.reject_ as (p1: Error) => void;
-          fbsAsync(toCall.bind(null, this.error_ as Error))();
+          const toCall = this._reject as (p1: FirebaseStorageError) => void;
+          fbsAsync(toCall.bind(null, this._error as FirebaseStorageError))();
           break;
         default:
           triggered = false;
           break;
       }
       if (triggered) {
-        this.resolve_ = null;
-        this.reject_ = null;
+        this._resolve = undefined;
+        this._reject = undefined;
       }
     }
   }
 
-  private notifyObserver_(observer: Observer<UploadTaskSnapshot>): void {
-    const externalState = taskStateFromInternalTaskState(this.state_);
+  private _notifyObserver(observer: Observer<UploadTaskSnapshot>): void {
+    const externalState = taskStateFromInternalTaskState(this._state);
     switch (externalState) {
       case TaskState.RUNNING:
       case TaskState.PAUSED:
@@ -541,39 +553,43 @@ export class UploadTask {
       case TaskState.CANCELED:
       case TaskState.ERROR:
         if (observer.error) {
-          fbsAsync(observer.error.bind(observer, this.error_ as Error))();
+          fbsAsync(
+            observer.error.bind(observer, this._error as FirebaseStorageError)
+          )();
         }
         break;
       default:
         // TODO(andysoto): assert(false);
         if (observer.error) {
-          fbsAsync(observer.error.bind(observer, this.error_ as Error))();
+          fbsAsync(
+            observer.error.bind(observer, this._error as FirebaseStorageError)
+          )();
         }
     }
   }
 
   /**
    * Resumes a paused task. Has no effect on a currently running or failed task.
-   * @return True if the operation took effect, false if ignored.
+   * @returns True if the operation took effect, false if ignored.
    */
   resume(): boolean {
     const valid =
-      this.state_ === InternalTaskState.PAUSED ||
-      this.state_ === InternalTaskState.PAUSING;
+      this._state === InternalTaskState.PAUSED ||
+      this._state === InternalTaskState.PAUSING;
     if (valid) {
-      this.transition_(InternalTaskState.RUNNING);
+      this._transition(InternalTaskState.RUNNING);
     }
     return valid;
   }
 
   /**
    * Pauses a currently running task. Has no effect on a paused or failed task.
-   * @return True if the operation took effect, false if ignored.
+   * @returns True if the operation took effect, false if ignored.
    */
   pause(): boolean {
-    const valid = this.state_ === InternalTaskState.RUNNING;
+    const valid = this._state === InternalTaskState.RUNNING;
     if (valid) {
-      this.transition_(InternalTaskState.PAUSING);
+      this._transition(InternalTaskState.PAUSING);
     }
     return valid;
   }
@@ -581,14 +597,14 @@ export class UploadTask {
   /**
    * Cancels a currently running or paused task. Has no effect on a complete or
    * failed task.
-   * @return True if the operation took effect, false if ignored.
+   * @returns True if the operation took effect, false if ignored.
    */
   cancel(): boolean {
     const valid =
-      this.state_ === InternalTaskState.RUNNING ||
-      this.state_ === InternalTaskState.PAUSING;
+      this._state === InternalTaskState.RUNNING ||
+      this._state === InternalTaskState.PAUSING;
     if (valid) {
-      this.transition_(InternalTaskState.CANCELING);
+      this._transition(InternalTaskState.CANCELING);
     }
     return valid;
   }
