@@ -17,7 +17,6 @@
 import { assert } from 'chai';
 import { FbsBlob } from '../../src/implementation/blob';
 import { Location } from '../../src/implementation/location';
-import { getMappings } from '../../src/implementation/metadata';
 import { Unsubscribe } from '../../src/implementation/observer';
 import { TaskEvent, TaskState } from '../../src/implementation/taskenums';
 import { Headers } from '../../src/implementation/xhrio';
@@ -26,12 +25,11 @@ import { StorageService } from '../../src/service';
 import { UploadTask } from '../../src/task';
 import { makePool, emptyAuthProvider } from './testshared';
 import { StringHeaders, TestingXhrIo } from './xhrio';
+import { FirebaseApp } from '@firebase/app-types';
 
 const testLocation = new Location('bucket', 'object');
 const smallBlob = new FbsBlob(new Blob(['a']));
 const bigBlob = new FbsBlob(new Blob([new ArrayBuffer(1024 * 1024)]));
-
-const mappings = getMappings();
 
 const fakeMetadata = '{ "downloadTokens": "a,b" }';
 
@@ -59,7 +57,11 @@ function storageServiceWithHandler(handler: RequestHandler): StorageService {
     xhrio.simulateResponse(response.status, response.body, response.headers);
   }
 
-  return new StorageService(null, emptyAuthProvider, makePool(newSend));
+  return new StorageService(
+    {} as FirebaseApp,
+    emptyAuthProvider,
+    makePool(newSend)
+  );
 }
 
 function fakeServerHandler(): RequestHandler {
@@ -191,15 +193,12 @@ describe('Firebase Storage > Upload Task', () => {
   it('Works for a small upload w/ an observer', done => {
     const storageService = storageServiceWithHandler(fakeServerHandler());
     const task = new UploadTask(
-      {} as Reference,
-      storageService,
-      testLocation,
-      mappings,
+      new Reference(storageService, testLocation),
       smallBlob
     );
     task.on(
       TaskEvent.STATE_CHANGED,
-      null,
+      undefined,
       () => assert.fail('Unexpected upload failure'),
       () => done()
     );
@@ -207,10 +206,7 @@ describe('Firebase Storage > Upload Task', () => {
   it('Works for a small upload w/ a promise', () => {
     const storageService = storageServiceWithHandler(fakeServerHandler());
     const task = new UploadTask(
-      {} as Reference,
-      storageService,
-      testLocation,
-      mappings,
+      new Reference(storageService, testLocation),
       smallBlob
     );
     return task.then(snapshot => {
@@ -220,10 +216,7 @@ describe('Firebase Storage > Upload Task', () => {
   it('Works for a small upload canceled w/ a promise', () => {
     const storageService = storageServiceWithHandler(fakeServerHandler());
     const task = new UploadTask(
-      {} as Reference,
-      storageService,
-      testLocation,
-      mappings,
+      new Reference(storageService, testLocation),
       smallBlob
     );
     const promise: Promise<string | null> = task.then<string | null>(
@@ -239,20 +232,27 @@ describe('Firebase Storage > Upload Task', () => {
   it('Works properly with multiple observers', () => {
     const storageService = storageServiceWithHandler(fakeServerHandler());
     const task = new UploadTask(
-      {} as Reference,
-      storageService,
-      testLocation,
-      mappings,
+      new Reference(storageService, testLocation),
       smallBlob
     );
 
     let badComplete = false;
-    const h1: Unsubscribe = task.on(TaskEvent.STATE_CHANGED, null, null, () => {
-      badComplete = true;
-    }) as Unsubscribe;
-    const h2: Unsubscribe = task.on(TaskEvent.STATE_CHANGED, null, null, () => {
-      badComplete = true;
-    }) as Unsubscribe;
+    const h1: Unsubscribe = task.on(
+      TaskEvent.STATE_CHANGED,
+      undefined,
+      undefined,
+      () => {
+        badComplete = true;
+      }
+    ) as Unsubscribe;
+    const h2: Unsubscribe = task.on(
+      TaskEvent.STATE_CHANGED,
+      undefined,
+      undefined,
+      () => {
+        badComplete = true;
+      }
+    ) as Unsubscribe;
 
     let resumed = 0;
 
@@ -270,8 +270,8 @@ describe('Firebase Storage > Upload Task', () => {
           }
           lastState = snapshot.state;
         },
-        null,
-        null
+        undefined,
+        undefined
       );
     })();
     h1();
@@ -280,7 +280,7 @@ describe('Firebase Storage > Upload Task', () => {
     return new Promise(resolve => {
       task.on(
         TaskEvent.STATE_CHANGED,
-        null,
+        undefined,
         () => {
           assert.fail('Upload failed');
         },
@@ -295,10 +295,7 @@ describe('Firebase Storage > Upload Task', () => {
   it("Works properly with an observer missing the 'next' method", () => {
     const storageService = storageServiceWithHandler(fakeServerHandler());
     const task = new UploadTask(
-      {} as Reference,
-      storageService,
-      testLocation,
-      mappings,
+      new Reference(storageService, testLocation),
       smallBlob
     );
     return new Promise(resolve => {
@@ -316,10 +313,7 @@ describe('Firebase Storage > Upload Task', () => {
   function runNormalUploadTest(blob: FbsBlob): Promise<void> {
     const storageService = storageServiceWithHandler(fakeServerHandler());
     const task = new UploadTask(
-      {} as Reference,
-      storageService,
-      testLocation,
-      mappings,
+      new Reference(storageService, testLocation),
       blob
     );
 
@@ -404,7 +398,7 @@ describe('Firebase Storage > Upload Task', () => {
 
     let completeTriggered = false;
 
-    task.on(TaskEvent.STATE_CHANGED, null, null, () => {
+    task.on(TaskEvent.STATE_CHANGED, undefined, undefined, () => {
       fixedAssertFalse(completeTriggered);
       completeTriggered = true;
 
@@ -438,10 +432,7 @@ describe('Firebase Storage > Upload Task', () => {
       fixedAssertTrue(lastIsAll);
 
       const task2 = new UploadTask(
-        {} as Reference,
-        storageService,
-        testLocation,
-        mappings,
+        new Reference(storageService, testLocation),
         blob
       );
       const events2: string[] = [];
